Moved `types::Second` and `types::Unit` to new `unit` module
Also review feedback.
