+++ src/hyperfine/format.rs
@@ -1,4 +1,4 @@
-use hyperfine::types::{Second, Unit};
+use hyperfine::units::{Second, Unit};
 
 /// Format the given duration as a string. The output-unit can be enforced by setting `unit` to
 /// `Some(target_unit)`. If `unit` is `None`, it will be determined automatically.
