+++ src/hyperfine/warnings.rs
@@ -2,7 +2,7 @@ use std::fmt;
 
 use hyperfine::format::format_duration;
 use hyperfine::internal::MIN_EXECUTION_TIME;
-use hyperfine::types::Second;
+use hyperfine::units::Second;
 
 /// A list of all possible warnings
 pub enum Warnings {
